Issue #81 & #74 message repository (#82)
* validator - domain - validator - MessageRepository

* validator - persistence - validator - memory - dummy MemoryMessageRepository

* validator - domain - validator - MessageRepository - make `add` a method

* memory-repository - extract generic impl to be used in MemoryRepositories

* memory-repository - remove ExistCmp

* Cargo.toml - add `memory-repository` to workspace

* memory-repository - add denies: `clippy` & `rust_2018_idioms`

* sentry - Remove `derive(Debug)` from ChannelResource & ChannelRepository

* memory - make `cmp` function `Arc` & `Send` + `Sync` for thread safety requirements

* sentry - use `memory-repository`

* memory-repository - impl `list_all` & cleanup `list`

* [test] memory-repository - list_all

* memory-repository:
- cmp_value instead of direct record search
- `find()` method

* sentry - persistence - channel - memory - refactor to use `MemoryRepository`

* validator - Cargo.toml - add `memory-repository`

* memory-repository - add comment for the `new` method

* domain - validator - message - `derive(Clone)`

* sentry - persistence - channel - memory - fix `clippy` warnings

* validator - domain - validator - MessageRepository add generic for State

* validator - persistence - validator - MemoryMessageRepository `add()` impl

* domain - validator - ValidatorId

* domain - validator - message:
- add constants for all types
- Message - impl `is_type` & `is_types`

* validator - domain - validator - `latest` method
- persistence - validator - MemoryMessageRepository - `latest` method
- persistence - validator - MemoryMessageRepository - `add` method fix
- persistence - validator - MemoryMessage struct

* domain - validator - add TODO for ValidatorDesc

* validator - persistence - validator - fix comment

+++ sentry/src/application/resource/channel.rs
@@ -12,7 +12,7 @@ use std::sync::Arc;
 mod channel_create;
 mod channel_list;
 
-#[derive(Clone, Debug)]
+#[derive(Clone)]
 pub struct ChannelResource {
     pub channel_list_limit: u32,
     pub channel_repository: Arc<dyn ChannelRepository>,
@@ -42,12 +42,12 @@ impl_web! {
 
 #[derive(Extract)]
 struct ChannelListQuery {
-    page: Option<u32>,
+    page: Option<u64>,
     validator: Option<String>,
 }
 
 impl ChannelListQuery {
-    pub fn page(&self) -> u32 {
+    pub fn page(&self) -> u64 {
         match self.page {
             Some(page) if page >= 1 => page,
             _ => 1,
